Code/requirements.txt
PyQt5
requests
cryptography 
django-sslserver
django
django-widget-tweaks
djangorestframework